pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lazygitlab"
version = "0.1.0"
description = "A TUI application for browsing and commenting on GitLab merge requests"
requires-python = ">=3.11"
dependencies = [
    "python-gitlab>=4.0,<5.0",
    "textual>=0.60,<1.0",
    "pygments>=2.17,<3.0",
]

[project.optional-dependencies]
dev-services = [
    "pytest-asyncio>=0.23,<1.0",
]

[project.scripts]
lazygitlab = "lazygitlab.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lazygitlab*"]

[tool.pytest.ini_options]
testpaths = ["lazygitlab"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["lazygitlab"]
omit = ["*/tests/*"]

[tool.coverage.report]
fail_under = 80

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.4,<1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "RUF"]
ignore = ["S101"]
